drivers/power: bq2597x_munch: eliminate floating point data type

+++ drivers/power/supply/ti/bq2597x_charger_munch.c
@@ -57,7 +57,7 @@ enum {
 	ADC_MAX_NUM,
 };
 
-static float sc8551_adc_lsb[] = {
+static int sc8551_adc_lsb[] = {
 	[ADC_IBUS]	= SC8551_IBUS_ADC_LSB,
 	[ADC_VBUS]	= SC8551_VBUS_ADC_LSB,
 	[ADC_VAC]	= SC8551_VAC_ADC_LSB,
@@ -1115,7 +1115,7 @@ static int bq2597x_get_adc_data(struct bq2597x *bq, int channel,  int *result)
 
 		if (bq->chip_vendor == SC8551) {
 			kernel_neon_begin();
-			*result = (int)(t * sc8551_adc_lsb[channel]);
+			*result = (t * sc8551_adc_lsb[channel]);
 			kernel_neon_end();
 		}
 	}
